Add config knob to suppress Trace logging
This mod, maybe due to its "beta" status, is VERY chatty.

Out of 15'000 lines of log, 14'000+ came from this mod.

This makes it VERY difficult to troubleshoot other mods.

So, I added a config.json knob to be able to temporarily disable Trace
logging from this mod. The knob defaults to "false", of course.